Finished bestTimeForStocks.py Completed the optimization of this program, and now it runs in average of 715ms when running through the extremely large test cases on leetcode.
I optimized like this
------------------------
Instead of my previous way of gathering the highestPrice a stock could have
made money through nesting a for loop to check all possibilitied, I realized
that I could set a variable to the next lowest value within that list, we can
attempt to find a higher stock increase gap from there. That works due to the fact
that we will only lower our lowCount value when the days stock is lower then the
current stock, so if it is higher we are increasing our highestPrice already.

This method makes it so that we can find the best day with only one iteration
of the prices list instead of a nested iteration.