Made If statements
Made these if statements to determine whether or not if I am at the bottom or top of the list, then whether or not if I am in a corner. However, my issue is I don't think this will work since I have to also have an understanding of what direction I am supposed to be going if I am in the middle somehow. So I am pushing this but going to keep working on it